Fix issue with program not terminating
Some classes like the vertex buffer clean up their opengl data when
destructing. The Pong class creates gameobjects and meshes which contain
these classes. Because of the scope the Pong class was not destructing
until after glfwTerminate() which caused an infinite loop and the
program would not terminate. Along with this, the Renderer class has a
static allocated Shader object. This also created problems with
program termination, so the Renderer::Cleanup funciton will free this
object