Only call for collision on enter. Guard against duplicate calls
OnCollision is used for an initial collision between two gameobjects.
Currently if they are still colliding afterwards it calls nothing, but
would call something like OnCollisionStay(). When one gameobject finds a
collision with another it calls OnCollision() on both. Prevent from
calling this again when the objects check for collision in the reverse
order