Remove unecessary vector keeping track of what collisions checked
Since it is a double for loop through the same list of objects, the
inner loop can start at the index after the outer. This way there is no
need to check for it being the current gameobject, or one that has
already been checked for a collision with. Also includes some cleanup