Add collision checking and create pong game loop
GameObjects can now create a ColliderBox that represents collidable
bounds of themselves. The Pong class in currently in charge of checking
for collision between objects. On every frame it will check every
GameObject pair currently active using its ColliderBox's if they are
present. If there is a collision it may call OnCollisionStart() if this
is the first collision of these objects, or OnCollisionStay() if they
were also colliding last frame. If they do not collide but did the frame
before, call OnCollisionEnd().