Add Timer::frameTime for frame independence
The Timer::frameTime variable measures the processing time from the
previous frame to this one. This variable is use in variable calulations
to accomplish things over time to make them framerate independent. Ex.
movement. A movement addition vector to position in OnUpdate() should
be multiplied by this value otherwise running OnUpdate() more times
will result in more additions to position in a second.