Introduce a callback timer class
This class creates a thread and will be able to take in timer requests
for a time and a function to callback. However the timer thread cannot
execute this function because it would likely cause data races with the
game loop. So these callbacks will be handed back to the game loop
thread when ready for execution.