Add timer callbacks for gameobjects functions and make game framerate independent(#9)
The Timer class will be able to take in timer requests
for a time and a function to callback. When the timer is finished, on the
next game loop, the function callback will be executed

This also introduces the Timer::frameTime variable. 
This variable measures the processing time from the
previous frame to this one. This variable is use in variable calulations
to accomplish things over time to make them framerate independent. Ex.
movement. A movement addition vector to position in OnUpdate() should
be multiplied by this value otherwise running OnUpdate() more times
will result in more additions to position in a second.