Create Real TIme Logger (#10)
This introduces a real time logger and real time logger thread. New Log() and RealTimeLog() calls will go to respective Log class handlers. Under the hood it uses a concurrent lock-free queue and libfmt to format the logs, then uses spdlog to format the date and log level. Real Time Logs are meant to be used anytime when working directly in the game loop to avoid locking/allocations that would slow things down.