Fix problems with incorrect textures and spacing
Any opengl objects now need to delete their copy constructors. This is
because the renderer id is used to free the object on the GPU side in
the destructor. If we copy an OpenGL object, it may be freed from under
us. So instead use move constructors only has trying to delete an object
with an id of 0 will not do anything.

And fix some calulations in text spacing using kern and left side
bearings