Create font and text rendering. Show player and opponent scores visually. (#12)
Much more complicated than I expected, this introduces a way to render text to the screen. Using stb_truetype to load in a true type font file (.ttf) it does many calculations to get the offset each character needs to have from the top of the line, spacing from other characters, and etc. It creates a btimap image that is then saved to a texture. To actually render, each character of the string of text is represented by a simple Rectangle quad. Then the bitmap image is overlayed onto that quad as the texture.

Now the score controller will dynamically create two text objects for each player's score. Upon scoring, the corresponding text object's text is changed, and the text object is essentially recalculated and recreated. Likely not the most efficient way to do this; by rereading the font file every time the text is changed.

Additionally, any opengl objects now need to delete their copy constructors. This is
because the renderer id is used to free the object on the GPU side in
the destructor. If we copy an OpenGL object, it may be freed from under
us. So instead use move constructors only has trying to delete an object
with an id of 0 will not do anything.