Implement sounds and audio mixing (#13)
A naive approach to adding sound output to the game. This uses portaudio for the audio callback. Sounds can be held on gameobjects then played through the PlaySound() function. This adds a new playing sound to the audio mixer. In each callback it reads from the audio file and mixes any together. Playing sounds just hold an index for the current sample it is reading, and a pointer to the Sound which holds the audio file data.

This also includes some code for spatializing the audio based on where the audio play position is. This is enabled via config, just to experiment a bit