Create classes for containing application window, and game engine. (#16)
A lot of this is moving around code and dealing with class structure.
Previously many of these systems were in main.cpp and are now into, ApplicationWindow which is meant to manage the actual window for whatever platform, and the Engine class which is meant to encapuslate the entire game "engine" as a whole. This also introduces the ability to dynamically resize the window without breaking the game. Before it would have been much more annoying to do.