Introduce Entity Component System (#17)
A pattern that is in most game engines is the Entity Component System. This pattern allows for the creation of unique gameobject/entities that have very specific behaviors, but may share behaviors between each other through the use of components. Instead of using inheritance to create game objects with meshes/colliders/transforms, this uses composition to attach all components to this game object.

Up until now the gameobject class had a mesh and collider as a protected variable, so any gameobjects like the player can create both the mesh and collider, while the score controller only creates a collider. Instead these classes override a InitalizeComponents() function where they AddComponent<Component>() with each component needed. The system part 
 of an ECS are like the collision manager. These systems operate on the relevant components.

A common feature when using an ecs, is trying to optimize for cache locality. When iterating over thousands of gameobjects, if the components are stored contiguously that can be significantly better than stored on the entity itself and access through the entity only. So instead of storing the components on the entity/gameobject, they are primarily allocated in a vector for each component. The systems usually do not care that the component is attached to an entity because everything it needs would be stored in the component.

I think there may be multiple parts to this. Right now, gameobjects hold components, while also supporting the OnStart and OnUpdate functions and behavior. Instead this should be a component itself and be attached to a purely empty gameobject. And the UI elements also need to be updated to use this system