Entity Component System: Behavior components and better scene creation (Part 2) (#18)
With the entity component system (ecs) introduced, gameobjects now are
composed of mulitple components to give them behavior. However the idea
of a game object having behavior through `OnStart()`, `OnUpdate()`, or
`OnCollision()` callbacks was tied to gameobjects themselves. For a
proper ecs, this should be in its own component `Behavior.` It is a
little bit trickier though because Behaviors are meant to be overridden
by users. And overriding the base Behavior class causes problems with
the CRTP Component chain. Instead of different behavior types being
stored in their own vector, they are stored in the Behavior vector. So
we need separate typing to be able to get out certain behaviors out from
the others.

Another big change here is how scenes are loaded. For now, a Scene class
will define a function to `BuildScene()` which creates the scene from
helper functions `NewObject()` and `NewObjectFromBlueprint()`. An object
Blueprint is nothing more than a way to quickly define multiple
instances of the same object. These objects are put into a vector and
handed of to the game loop to be properly loaded.