Change UI Elements to using components
Changes buttons, sliders, checkboxes, and texts to components that can
be added to game objects. A few different things needed to be changed.
The bounds in colliderboxes needed to be brought out to a smaller object
that just contains the bounds and bounds checking. This is because
components like button should not have other components. This also means
the mesh component needed to have the new "Graphic" extracted out of it.
The renderer now knows how to render generic graphics rather than any
mesh