Create font strings on intialization instead of during runtime (#20)
Prevously if a text component was changed/created it would open the font
file, read it, and create the character images using stb true type. This
can be costly to do over and over each frame the text is changed.

Instead `Font` objects are created at initalization and stored within
the engine. It contains most the details for each font's character
including the bitmap image of the character, its offsets, and sizing
data. This way when recomputing the text component, it only needs to add
some offsets together to get the new text data.

Previously measured loadtimes for the 3 mains scene and their new load
times are as follows
Title       100-115ms -> 1-5ms
Settings 285-300ms -> 6-8ms
Game     39-40ms     -> 0-1ms
So a massive improvement! Of course there is a slight tradeoff that the
initalization of the engine will take longer to complete.

Some things to note/ignore:
A lot of this is moving computations from text.cpp to font.cpp
Changes to scene files were made to adjust to new default scaling of
fonts

The main idea is that the `FontBank` will load all fonts present in the
config and store them in a map where a font name maps to a new font
object. Text components will look for this font object, and give it the
text it wants to create a component from. The font object will give back
a `FontString` along with a vector of `FontCharacters`. Using the data
from these objects, each text character can be made.

Another smallish change is the use of a new namespace `image` Instead of
Texture holding the methods `ConvertAlphaImageToRGBA` and
`FlipImageVertically` statically, they are free functions under the
namespace `image`