Add in a gameover screen
Game will transition to a gameover screen when the opponent or player
reaches the target score set via config.

In order for this to happen the score controller will persist past the
game screen to gameover. The gameover controller will retrieve the
player and opponent score in order to set the text accordingly