Remove most singletons from engine creation. (#22)
A little hard to review since a lot of this is removing static keywords,
removing GetInstance() calls, and removing static functions. The main
idea is that in order to remove singletons, areas of the code that need
to be accessed use free functions under a specific namespace. Ex to
change the window's title from a user defined `Behavior`,
`application::SetWindowTItle("My Game");` is used.

Some of the singletons being removed:
Engine, Renderer, Input, Pong, ApplicationWindow

And importantly this uses depency injection to build the thing the game
engine needs. The Engine class remains the top level class holding
everything else `Renderer`, `ApplicationWindow`, `Input`, `Pong`,
`Input` does depend on both the renderer and application window to
properly use callbacks so they are passed into the constructor.